pull out all test code from non-test source
